shabeeh/requirements.txt
pandas
google-cloud-storage
sqlalchemy
psycopg2-binary
